pull out key to update into its own file and document
